rational.*: added a == and != operator ro the RationalNTL class. testEhrhart/*: added options to make a random-rational vertex polytope. (I force the randome polytope to have at least one integer vertex, so the number of lattice points is positive).
valuation/*: added an automatic testing function to find volums on different polytopes. I do call polymake to find the volume but this number is meaning less if the polytope is not fully dim, and so I do not use it to check our answers. Instead, I simply check that both of our methods [Lawrence & triang.] return the same answer for the volume. I'm also now calling free on some data structures. It seems that some of the variables have pointers to the same objects, so you cannot simply just call free on everything. I need to look closely at what *matrix, *temp, etc are pointing to.

	valuation/Driver.cpp: The main function calles 3 other functions, but only one should be compiled at a time. 1) the old main programe, 2) test a random polytope (usage: ./ValuationComputation amb_dim num_points), and 3) A automatic test script.


ok, good night.