rational.*: added more overloaded operators. residue, newIntegration.*: added functions to compute the integral over rational-vertex simplex. valuation, PolytopeValuation, Driver.* changed the driver and timed test the integration by dilation and the new "direct" method.
It turns out the "direct" method of integration is slower than the dilation method, even after trying to put off reducing most of the fractional results. The dilation method was faster for hyper-rectangles and for random polytopes (made from polymake).

I am commiting this direct integration method, and then I will merge/revert the trunk back to 1221 so there is a record of the direct integration method... but it should not be used.