added BirkhoffGenerator.mpl
testingSL_lib: updated the main volume function. Currently, we k tests using latte/old maple, and then do k tests with the SL functions. My goal is to mix the order of which method is used, so that that you do not have to wait for every test before you compair the results!

testingSLVolume.mpl: Added profile recording and printing for the SL functons.