Added an automatic test for the top 3 coef.
Running table_time_sl_ehrhart(2, 50, file) will find the average time it takes to find the top 3 ehrhart coeff. with 50 test samples for each dimention between 2 and infinity. When it takes longer than 30min to find the top 3 coeff. of one simplex, the script stops. The final and partial results are saved in a file.