renamed the testEhrhart folder to buildPolytopes. Will need to work on making the interface to the polytope building classes nicer.
moved the testing functions in valuation to valuation/testCorrectnessAndBenchmark. We sould place the functions that we will write for the paper in this folder. 

valuation/valuation.*, valuation/Driver.cpp: removed the volume testing functions and placed them in the testCorrestnessAndBenchmark folder.

Cleaned up the makefile to reflect the above changes.