Perturbation: print out the number of times we did not correctly find a perturbation.
added exe to integrate all the .latte files in a directory and save their results to a log file. Later will will have to parse this log file and do stats.

buildDatabaseRandomPolytopesDriver: you do not have to add a / at the end of the directory names now. Also we are making random polytopes with 2*dim +1 points currently.