rational.*: added a "to string" fuction.
IntegrationDB: moved a function in the driver to the class.
BuildPolytope: now centering the vertices before finding the dual. This is still not correctly giving me the dual. After looking into this more, I NOT know how to find the dual in polymake. 

This is not the final commit, just saving work. It does compile.