Perturbation.*: now picking random point on the curve (t,t^2,t^3,...)
Database stuff: adding/and testing cases.

added fixDualPolytopesInDatabaseDriver. If you did not make dual's, this will go back and make them. dual polymake fies are deleted (not needed)

added printDatabaseTimeTablesDriver: makes a table of the times for everything per dim. Currently does this in html, but will need to change this to latex (...I need to relearn how to do it in latex).