Makefile:: added the .cpp/.h files from the stokes integration method.
This is a major update to a lot of core reading/preprocess files, and new functions for the stokes recursive valuation method have been added.

The stokes method is not finished, and it does not work. Also, the interface to the stokes method is not done. For example, calling the stokes method is 'hidden' in the valuation driver, there are many debug-statements, few comments, and I have cut-and-pasted core latte reading/preprocessing function and modified them for the stokes method. A better solution is to divide up the core latte functions and only call the ones I need. I will do all this once the THEORY of how the stokes method should work is known (we do not know yet the correct formula for the recursive valuation for non-full dimentional polytopes.)


This is commint 1/2. (I want a new revision number starting here just in case the stokes update messing something up).