this is commit 2/2.
This contains the unfinished stokes method code with a poor interface and poor code reuse (I cut/pasted the core latte read function instead of dividing them because I didn't want to mess anything up. This will be fixed once the stokes method is done.)

So said, again, these last two commits is me saving my work. But the user has no knowlege the stokes method is half there...it is save to take this as the next release to latte.