banner: changed the name of the software...again.
valuation.*: added command line options for interactive mode and the unweighted case for the top ehrhart mode.

latteException.h: new comments.

Conebycone....mpl: added a new function that computes the coefficients one at a time. 
We now compute the top k coefficients and not the top k+1.
We are using tools/gensym for one of the variables.
We can print the polynomial to a file
todo:
--comments
--remove my old function and the "just simplex" functions. Remove the old latte interface functions
--run a few tests to check correctness. Maybe do this via a script? 


general todo list:
--update the manual.